requirements.txt
numpy
redis
kafka-python
lz4
blosc
ffmpeg-python
natsort
attrdict
PyTurboJPEG
opencv-python